added signup endpoint

+++ internal/app/models/models.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type User struct {
-	ID          int64     `json:"id"`
+	ID          int       `json:"id"`
 	Email       string    `json:"email"`
 	Username    string    `json:"username"`
 	Password    string    `json:"password"`
@@ -12,8 +12,8 @@ type User struct {
 }
 
 type Post struct {
-	ID          int64      `json:"id"`
-	UserID      int64      `json:"user_id"`
+	ID          int        `json:"id"`
+	UserID      int        `json:"user_id"`
 	Title       string     `json:"title"`
 	Content     string     `json:"content"`
 	Likes       int        `json:"likes"`
@@ -24,9 +24,9 @@ type Post struct {
 }
 
 type Comment struct {
-	ID          int64     `json:"id"`
-	UserID      int64     `json:"user_id"`
-	PostID      int64     `json:"post_id"`
+	ID          int       `json:"id"`
+	UserID      int       `json:"user_id"`
+	PostID      int       `json:"post_id"`
 	Content     string    `json:"content"`
 	Likes       int       `json:"likes"`
 	Dislikes    int       `json:"dislikes"`
